moved NSException workaround to separate file
