+++ SourceryRuntime/Sources/AST/AssociatedType.swift
@@ -1,17 +1,5 @@
 import Foundation
 
-#if !canImport(ObjectiveC)
-public class NSException {
-    static func raise(_ name: String, format: String, arguments: CVaListPointer) {
-        fatalError ("\(name) exception: \(NSString(format: format, arguments: arguments))")
-    }
-
-    static func raise(_ name: String) {
-        fatalError("\(name) exception")
-    }
-}
-#endif
-
 /// Describes Swift AssociatedType
 #if canImport(ObjectiveC)
 @objcMembers
